Livened up the Readme
The previous readme left a bit to be desired.  It didn't have a
bi-directional example, it was organized a little weird, and you called
using beefy a "hack".

Here I've rearranged most of the content, introduced with an even more
limited no-context example to get people excited, and then included a
just slightly modified full `example` file.

Fixes #10